Implement Decode Block 4A (Date and Time) save difference between local and rdsTime Implement static method for getting the current calculated rdsTime in rdsDecoder, if time not available the utc of the maschine is used (initial value) use rdsTime for rdsDecoder for tmcMsg Fix some debug outputs

+++ bitsetqueue.cpp
@@ -10,7 +10,7 @@ bitsetqueue::~bitsetqueue() {
 
 }
 
-void bitsetqueue::addValue(int value, const short length) {
+void bitsetqueue::addValue(int16_t value, const short length) {
 	int i = length - 1;
 	for(;i>-1;i--){
 		bool bit = value & (1 << i);
@@ -26,8 +26,8 @@ void bitsetqueue::setBit(bool value){
 	bits.set(size++, value);
 }
 
-int16_t bitsetqueue::getNextBits(short size) {
-	int16_t returnInt = 0;
+int32_t bitsetqueue::getNextBits(short size) {
+	int32_t returnInt = 0;
 
 
 	int i;
